source_code/requirements.txt
Django==1.11.21
django-background-tasks==1.2.0
django-compat==1.0.15
djangorestframework==3.9.4
numpy
gitpython
guesslang
pymongo
